Move snapshot example for tests to snapshot_list_test.go
