+++ mcc/odin/odin/odin_test.go
@@ -3,7 +3,6 @@ package odin_test
 import (
 	"testing"
 
-	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/rds"
 	"github.com/go-test/deep"
 
@@ -42,19 +41,6 @@ func (tc *testCase) check(actual interface{}, err error, t *testing.T) {
 			)
 		}
 	}
-}
-
-var exampleSnapshot1Type = aws.String("db.m1.medium")
-var exampleSnapshot1DBID = aws.String("production-rds")
-var exampleSnapshot1ID = aws.String("rds:production-2015-06-11")
-
-var exampleSnapshot1 = &rds.DBSnapshot{
-	AllocatedStorage:     aws.Int64(10),
-	AvailabilityZone:     aws.String("us-east-1c"),
-	DBInstanceIdentifier: exampleSnapshot1DBID,
-	DBSnapshotIdentifier: exampleSnapshot1ID,
-	MasterUsername:       aws.String("owner"),
-	Status:               aws.String("available"),
 }
 
 type getLastSnapshotCase struct {
