+++ mcc/odin/odin/snapshot_list_test.go
@@ -1,0 +1,19 @@
+package odin_test
+
+import (
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/rds"
+)
+
+var exampleSnapshot1Type = aws.String("db.m1.medium")
+var exampleSnapshot1DBID = aws.String("production-rds")
+var exampleSnapshot1ID = aws.String("rds:production-2015-06-11")
+
+var exampleSnapshot1 = &rds.DBSnapshot{
+	AllocatedStorage:     aws.Int64(10),
+	AvailabilityZone:     aws.String("us-east-1c"),
+	DBInstanceIdentifier: exampleSnapshot1DBID,
+	DBSnapshotIdentifier: exampleSnapshot1ID,
+	MasterUsername:       aws.String("owner"),
+	Status:               aws.String("available"),
+}
